New logger, minor bugfixes and refactoring
1. A new logging engine. Currently it can be configured via environment variables:

   - FAR.LOG.LEVEL:
     The log level. Can be one of the following values: "off", "fatal", "error", "warning", "notice", "info", "debug", "trace", "all".
     Default value: off

   - FAR.LOG.SINK:
     Log sink(s). Can contain one or more comma- or semicolon-separated values:
      - "debug": the system debugger.
      - "console": a new buffer in the same console window, available via the Ctrl+Shift+L hotkey or the "far:log" command.
      - "file": a text file with a unique name. The path to the file can be specified via FAR.LOG.SINK.FILE.PATH.
      - "pipe": similar to "console", but in a new window.

2. syslog.* and everything related to it removed.

3. Plugins cache:
   - Removed replacing of '\' with '/' (remnants of the Windows Registry days).
   - Only lower 32 bits of timestamps were checked (with potential collisions every ~7 minutes).

4. Skipping the file size in descriptions didn't work.

5. Added logging in quite a few places where it should be, more to come.

6. Minor refactoring related to #1.

+++ far/memcheck.cpp
@@ -284,7 +284,7 @@ private:
 				L"\nWide: "sv, printable_wide_string(UserAddress, std::min(BlockSize, Width * sizeof(wchar_t))),
 				L"\nStack:\n"sv);
 
-			DWORD64 Stack[ARRAYSIZE(MEMINFO::Stack)];
+			uintptr_t Stack[ARRAYSIZE(MEMINFO::Stack)];
 			size_t StackSize;
 
 			for (StackSize = 0; StackSize != std::size(Stack) && i->Stack[StackSize]; ++StackSize)
